// src/types/Pacientes.ts
export type Paciente = {
    id: number;
    nome: string;
    rg: number;
    cpf: number;
    sus: number;
    dataNascimento: string;
    mae: string;
    pai: string;
    naturalidade: string;
    profissao: string;
    rua: string;
    numero: number;
    bairro: string;
    complemento: string;
    cidade: string;
    estado: string;
    celular: string;
    email: string;
    senha: string;
    confirmarSenha: string;
}